requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
streamlit
langchain_groq
python-dotenv==1.0.0
selenium==4.12.0
playwright==1.38.0
html2text==2020.1.16
requests==2.31.0
crewai==0.28.8
langchain-core
bs4==0.0.1
beautifulsoup4==4.12.2
webdriver-manager==4.0.0
pydantic==2.3.0